Move classes in ChartIOHelper to their own files
